shell: Reorganise files used when invoking the shell
Organise shell files assuming that the shell used is invoked as /bin/sh.
Then, the files /etc/profile and ~/.profile will be read in that order.
If the environment variable ENV is set, then the file named in ENV will
be read.

The variable ENV is set to ~/.config/shell/shinit in ~/.profile. The
former file acts as a general initialisation file for interactive
shells. If it is detected that a particular shell is used (e.g., bash,
zsh, ksh), then the particular initialisation file for that shell,
according to the directory structure of this repository, will be
sourced. For example, if bash is used, then ~/.config/bash/bashrc will
be loaded. If no paricular shell is detected, ~/.config/shell/shinit
will configure the shell in a POSIX-compliant way.

The configurartion scripts in ~/.config/shell/shinit.d are meant to be
as portable as possible so that they can be sourced from other shell
specific files. In case more functionality or customisation is needed
for a particular shell, drop-in replacements can be used within that
shell directory. For example, prompt customisation in bash might use \[
and \] for delimiting non-printable characters. Then, a prompt
customisation script can be placed in ~/.config/bash/bashrc.d and
sourced from ~/.config/bash/bashrc while sourcing the remaining scripts
in ~/.config/shell/shinit.d.